% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_glmnet.R
\name{cv_glmnet}
\alias{cv_glmnet}
\title{Tuning parameter selection by cross-validation}
\usage{
cv_glmnet(X, Y, foldid, alpha = 1, error = "cox", offset = rep(0,
  nrow(X)), setlambda = F, lambda = 0, ...)
}
\arguments{
\item{X}{The matrix of all predictors. Each row represents a subject 
and each column represents a feature.}

\item{Y}{The survival time, represented by a \code{Surv} object.}

\item{foldid}{A vector of integers between 1 and \code{k} indicating  what fold each subject is in, where \code{k} is the number of folds.}

\item{alpha}{The second tuning parameter in elastic net (\eqn{\alpha}).}

\item{error}{The choice of cross-validation error. Set \code{error="cox"} for deviance under the Cox model and \code{error="Cindex"} for the C-index.}

\item{offset}{A vector of offset terms.}

\item{setlambda}{If \code{setlambda=TRUE}, then a user-supplied vector of tuning parameters (\eqn{\lambda}) would be used; otherwise, the sequence would be chosen automatically.}

\item{lambda}{The user-supplied sequence of tuning parameters (\eqn{\lambda}).}

\item{...}{Other arguments that can be passed to \code{glmnet}.}
}
\value{
A list containing results of the cross-validation fit:
   \item{cvm}{a vector of mean cross-validation errors over lambda values.}
   \item{lambda}{a vector of tuning parameter values.}
   \item{df}{a vector of mean numbers of variables selected over lambda values.}
   \item{min.cvm}{minimum value of mean cross-validation error.}
   \item{lambda.min}{the lambda value corresponding to the minimum (mean) cross-validation error.}
   \item{min.pos}{the position of the minimum (mean) cross-validation error in \code{cvm}.}
   \item{all.beta}{a list of estimated regression parameters over the folds.}
   \item{all.cvm}{a list of vectors of cross-validation errors over the folds.}
}
\description{
Performs k-fold cross-validation for the first tuning parameter
   of elastic net (lambda) in the same way as \code{cv.glmnet} in the package \code{glmnet} and outputs the selected lambda and the cross-validation errors for individual folds.
}
\seealso{
\code{glmnet}
}
