% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_glmnet.R
\name{cv_glmnet}
\alias{cv_glmnet}
\title{Tuning parameter selection by cross-validation}
\usage{
cv_glmnet(X, Y, foldid, alpha = 1, error = "cox", offset = rep(0,
  nrow(X)), setlambda = F, lambda = 0, ...)
}
\arguments{
\item{X}{The matrix of all predictors. Each row represents a subject (with total of n rows) and each column represents a feature (with total of d columns). Each column of X should be standardized.}

\item{Y}{The survival time, represented by a Surv object.}

\item{foldid}{A vector of integers between 1 and \code{k} indicating  what fold each subject is in, where \code{k} is the number of folds.}

\item{alpha}{The second parameter in elastic net (\eqn{\alpha}).}

\item{error}{The choice of cross-validation error. \code{error="cox"} for deviance under the Cox model and \code{error="Cindex"} for the C-index.}

\item{offset}{A vector of offset terms.}

\item{setlambda}{If \code{setlambda=TRUE}, then a user-supplied vector of tuning parameters (\eqn{\lambda}) would be used; otherwise, the sequence would be chosen automatically.}

\item{lambda}{The user-supplied sequence of tuning parameters (\eqn{\lambda}).}

\item{...}{Other arguments that can be passed to \code{glmnet}.}
}
\description{
Performs k-fold cross-validation for glmnet in the same way as \code{cv.glmnet} in the package \code{glmnet} and outputs the cross-validation error for each fold.
}
\seealso{
\code{glmnet}
}
